requirement.txt
streamlit==1.25.0
google-generative-ai==0.2.1
